feat: v1.0.0 — native C++ runtime, Gemma-4-12B coherent on NPU
Bring the native C++ runtime to a working release: Gemma-4-12B runs
coherently end-to-end on the XDNA2 NPU with no Python in the decode loop,
greedy-matching the Python runtime and llama.cpp token-for-token.

Runtime (runtime/cpp):
- Native XRT kernel registry, decode loop, OpenAI HTTP server (cpp-httplib).
- LM head tiled onto the NPU as quantized GEMVs (was a CPU-only fallback,
  and previously segfaulted reading the packed Q4_0 head as dense bf16).
- Self-contained byte-level BPE tokenizer (GemmaTokenizer) parsing a HF
  tokenizer.json, plus the Gemma chat template; replaces the byte stub.
- Correctness fixes: Gemma-4 per-layer output scale, float16 embedding
  decode, sqrt(hidden) embedding scale, greedy on raw (un-softcapped) logits.

Kernels:
- Fused FFN kernel (gate + up + GeGLU + down) with fp32 accumulation; bf16
  accumulation compounded to ~13% error over 48 layers and produced garbage.
  fp32 output accumulator fits tile memory via a 2-pass H-output split.
- Force the AIE design CLI to the full-column device (from_name n_cols=None)
  so multi-core kernels place (default n_cols=1 = single column).

Docs: README, ROADMAP (M11/M12), CHANGELOG, handoff note.

+++ kernels/ffn_fused/ffn_compute_activation.cc
@@ -6,8 +6,8 @@
 #define DIM_M 32
 #endif
 
-extern bfloat16 gate_accum[DIM_M];
-extern bfloat16 up_accum[DIM_M];
+extern float gate_accum[DIM_M];
+extern float up_accum[DIM_M];
 extern bfloat16 act[DIM_M];
 
 // Stable fast exponential approximation
